BIG UPDATE: DatabasePluginMixin
For the longest time I have trouble figuring out how to extend the
__unload for cogs with databases. This came to be because I needed to
serialize namedtuple, and being the sucky format it is,
JSON doesn't allow that. To make matters even worse, the __unload method,
which start with double underscore, gets name-mangled by python,
and discord.py detects ONLY the name-mangled __unload.

This was the best solution I could think of.

Signed-off-by: Ikusaba-san <Ikusaba-san@users.noreply.github.com>