Finally put config.json to good use
Unfortunately the default_help (if help is called with no arguments) can only be a string as of now
That might change if I can find a work around.
Also I'm really not liking the new converters in discord.py, it's way more verbose than needed.

In the meantime, enjoy some properties.