removed cog-specific prefixes
This is done because of a few reasons, namely:
- The code used to detect it was extremely ugly
- Using it was extremely unwieldy, as prefixes can't be full words,
  and command names can't have symbols in them.
- It doesn't really do anything, because all the commands are in the same namespace anyway,
  so two commands with the same name but different prefixes fall in the same command.

The replacement will be a global prefix, that can be different per server.
Doing this will allow for other things, such as:
- word prefixes
- custom command aliases
- using symbols in command names

Also, prefixes can now be placed in the help. Just specify {prefix} in the docstring of the command callback.
Along with a few other changes, such as inanyrole and inallrole for meta.py,
and command.all_recursive_names to list *all* possible qualified names for a command
If there are more readable/efficient ways of going about doing it let me know.