added id for welcome/leave messages; made group a toggle command
The latter is a breaking change. Before, invoking ->welcome or ->byebye on their own
would show the message that Chiaki would say. Now it will simply toggle whether
or not she would announce the message. This was because I needed a cleaner way
to disable welcome/leave messages, without deleting the whole message.

The big downside is that because now the group is technically a command,
all it's subcommands now need to have the same check. This is massive repetition,
and I have no idea how to reduce it, if at all.