Completely removed the custom checks in favour of the built-in ones
This is the largest breaking change I've done in a while. Now commands
simply require discord permissions, rather than having a admin/mod role or permissions.

The rationale was to simplify using Chiaki ten-fold, along with using some serious
hacks to get the actual requirements. This should make it a lot easier to use
some of the moderation features for Chiaki.

(Some of the files have been completely rewritten due to messing up the line endings)