Allow negative numbers for ->reason and ->case
In order to do this I have to cache the function that gets the current case number.
This can probably be optimized by doing an http req to get the latest case message.
but I don't wanna push it.

However, this does have to be optimized intensely, because the reason command
takes 0.7s which takes far too long for a simple command.