[feature?] Add custom trivia and use OTDB
I'll probably have to find a new way to get input for OTDB questions
to avoid people just spamming guesses.

Also the questions for diepio had to be changed in order to support using
a DB (which only takes in one answer). Aliases for answers might be
possible by using foreign keys, similar to what I did for mod-log targets,
but it will be a nightmare to copy.